services/linker/requirements.txt
nats-py>=2.13.1
